requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
aws-cdk-lib==2.187.0
aws-cdk.asset-awscli-v1==2.2.230
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==41.2.0
boto3==1.37.26
botocore==1.37.26
cattrs==24.1.3
click==8.1.8
colorama==0.4.6
constructs==10.4.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.111.0
publication==0.0.3
pydantic==2.11.2
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typeguard==2.13.3
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
